refactor(config): split project config modules
